#32: removed code smells
