+++ backend/src/main/kotlin/org/darkSolace/muse/ElysianMuseApplication.kt
@@ -1,0 +1,10 @@
+package org.darkSolace.muse
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication class ElysianMuseApplication
+
+fun main() {
+    runApplication<ElysianMuseApplication>()
+}
