+++ backend/src/main/kotlin/org/darkSolace/muse/layoutModule/model/FontFamily.kt
@@ -1,0 +1,6 @@
+package org.darkSolace.muse.layoutModule.model
+
+enum class FontFamily {
+    SANS,
+    SERIF
+}
