+++ backend/src/main/kotlin/org/darkSolace/muse/layoutModule/model/FontSize.kt
@@ -1,0 +1,7 @@
+package org.darkSolace.muse.layoutModule.model
+
+enum class FontSize {
+    SMALL,
+    MEDIUM,
+    LARGE
+}
