+++ backend/src/main/kotlin/org/darkSolace/muse/layoutModule/model/Layout.kt
@@ -1,0 +1,10 @@
+package org.darkSolace.muse.layoutModule.model
+
+import javax.persistence.Entity
+import javax.persistence.Id
+
+@Entity
+data class Layout(
+    @Id
+    var id: Long? = null
+)
