+++ backend/src/main/kotlin/org/darkSolace/muse/storyModule/model/Rating.kt
@@ -1,0 +1,10 @@
+package org.darkSolace.muse.storyModule.model
+
+enum class Rating {
+    GENERAL_AUDIENCE,
+    PARENTAL_GUIDANCE,
+    PARENTAL_GUIDANCE_13,
+    RESTRICTED,
+    NC_17,
+    ADULT_ONLY
+}
