+++ backend/src/main/kotlin/org/darkSolace/muse/userModule/model/Role.kt
@@ -1,0 +1,8 @@
+package org.darkSolace.muse.userModule.model
+
+enum class Role {
+    ADMINISTRATOR,
+    MEMBER,
+    MODERATOR,
+    SUSPENDED
+}
