+++ backend/src/main/kotlin/org/darkSolace/muse/userModule/model/UserTags.kt
@@ -1,0 +1,10 @@
+package org.darkSolace.muse.userModule.model
+
+enum class UserTags {
+    ARTIST,
+    ARTIST_INACTIVE,
+    AUTHOR,
+    BETA,
+    BETA_INACTIVE,
+    COMMENTER
+}
