+++ backend/src/main/kotlin/org/darkSolace/muse/userModule/repository/AvatarRepository.kt
@@ -1,0 +1,8 @@
+package org.darkSolace.muse.userModule.repository
+
+import org.darkSolace.muse.userModule.model.Avatar
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface AvatarRepository : CrudRepository<Avatar, Long>
