+++ backend/src/main/kotlin/org/darkSolace/muse/userModule/repository/UserRepository.kt
@@ -1,0 +1,12 @@
+package org.darkSolace.muse.userModule.repository
+
+import org.darkSolace.muse.userModule.model.Role
+import org.darkSolace.muse.userModule.model.User
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface UserRepository : CrudRepository<User, Long> {
+    fun findByUsername(username: String): User
+    fun findAllByRole(role: Role): List<User>
+}
