+++ backend/src/main/kotlin/org/darkSolace/muse/userModule/repository/UserSettingsRepository.kt
@@ -1,0 +1,8 @@
+package org.darkSolace.muse.userModule.repository
+
+import org.darkSolace.muse.userModule.model.UserSettings
+import org.springframework.data.repository.CrudRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface UserSettingsRepository : CrudRepository<UserSettings, Long>
