--- backend/src/main/kotlin/org/dark_solace/muse/ElysianMuseApplication.kt
@@ -1,10 +1,0 @@
-package org.dark_solace.muse
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication class ElysianMuseApplication
-
-fun main(args: Array<String>) {
-    runApplication<ElysianMuseApplication>(*args)
-}
