--- backend/src/main/kotlin/org/dark_solace/muse/layoutModule/model/FontFamily.kt
@@ -1,6 +1,0 @@
-package org.dark_solace.muse.layoutModule.model
-
-enum class FontFamily {
-    SANS,
-    SERIF
-}
