--- backend/src/main/kotlin/org/dark_solace/muse/layoutModule/model/FontSize.kt
@@ -1,7 +1,0 @@
-package org.dark_solace.muse.layoutModule.model
-
-enum class FontSize {
-    SMALL,
-    MEDIUM,
-    LARGE
-}
