--- backend/src/main/kotlin/org/dark_solace/muse/layoutModule/model/Layout.kt
@@ -1,10 +1,0 @@
-package org.dark_solace.muse.layoutModule.model
-
-import javax.persistence.Entity
-import javax.persistence.Id
-
-@Entity
-data class Layout(
-    @Id
-    var id: Long? = null
-)
