--- backend/src/main/kotlin/org/dark_solace/muse/storyModule/model/Rating.kt
@@ -1,10 +1,0 @@
-package org.dark_solace.muse.storyModule.model
-
-enum class Rating {
-    GENERAL_AUDIENCE,
-    PARENTAL_GUIDANCE,
-    PARENTAL_GUIDANCE_13,
-    RESTRICTED,
-    NC_17,
-    ADULT_ONLY
-}
