--- backend/src/main/kotlin/org/dark_solace/muse/userModule/model/Role.kt
@@ -1,8 +1,0 @@
-package org.dark_solace.muse.userModule.model
-
-enum class Role {
-    ADMINISTRATOR,
-    MEMBER,
-    MODERATOR,
-    SUSPENDED
-}
