--- backend/src/main/kotlin/org/dark_solace/muse/userModule/model/UserTags.kt
@@ -1,10 +1,0 @@
-package org.dark_solace.muse.userModule.model
-
-enum class UserTags {
-    ARTIST,
-    ARTIST_INACTIVE,
-    AUTHOR,
-    BETA,
-    BETA_INACTIVE,
-    COMMENTER
-}
