--- backend/src/main/kotlin/org/dark_solace/muse/userModule/repository/AvatarRepository.kt
@@ -1,8 +1,0 @@
-package org.dark_solace.muse.userModule.repository
-
-import org.dark_solace.muse.userModule.model.Avatar
-import org.springframework.data.repository.CrudRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface AvatarRepository : CrudRepository<Avatar, Long>
