--- backend/src/main/kotlin/org/dark_solace/muse/userModule/repository/UserSettingsRepository.kt
@@ -1,8 +1,0 @@
-package org.dark_solace.muse.userModule.repository
-
-import org.dark_solace.muse.userModule.model.UserSettings
-import org.springframework.data.repository.CrudRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface UserSettingsRepository : CrudRepository<UserSettings, Long>
